# Conflicts:
#	docker-compose.yml

+++ src/main/java/org/labonnefranquette/data/controller/CommandeController.java
@@ -5,8 +5,8 @@ import org.labonnefranquette.data.dto.impl.CommandeReadDTO;
 import org.labonnefranquette.data.exception.PriceException;
 import org.labonnefranquette.data.model.Commande;
 import org.labonnefranquette.data.model.enums.StatusCommande;
-import org.labonnefranquette.data.projection.CommandeListeProjection;
 import org.labonnefranquette.data.services.CommandeService;
+import org.labonnefranquette.data.utils.ControlInputTool;
 import org.labonnefranquette.data.utils.DtoTools;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,11 +14,15 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.web.bind.annotation.*;
 
+import jakarta.validation.Valid;
+
+import org.springframework.validation.annotation.Validated;
+
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/commandes")
+@Validated
 public class CommandeController {
 
     private final SimpMessagingTemplate template;
@@ -31,16 +35,23 @@ public class CommandeController {
     private CommandeService commandeService;
     @Autowired
     private DtoTools dtoTools;
-
+/*
+Les commandes ne sont pas récupérés par un endpoint REST mais par un websocket
     @GetMapping
     public ResponseEntity<List<CommandeReadDTO>> fetchAllCommandes() {
         List<Commande> commandes = commandeService.findAllCommande();
         List<CommandeReadDTO> commandesDtos = commandes.stream().map(commande -> dtoTools.convertToDto(commande, CommandeReadDTO.class)).toList();
         return new ResponseEntity<>(commandesDtos, HttpStatus.OK);
     }
-    @GetMapping("/status/{status}")
+*/
+
+    //Utilisé lors d'affichage de l'écran de la cuisine
+    @GetMapping(value= "/status/{status}", produces = "application/json")
     public ResponseEntity<List<CommandeReadDTO>> fetchAllCommandesEnCours(@PathVariable String status) {
         try {
+            if (!ControlInputTool.isValidString(status)) {
+                return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+            }
             StatusCommande statusCommande = StatusCommande.valueOf(status.replace("-","_").toUpperCase());
             List<Commande> commandes = commandeService.findAllCommandeWithStatut(statusCommande);
             List<CommandeReadDTO> commandesDtos = commandes.stream().map(commande -> dtoTools.convertToDto(commande, CommandeReadDTO.class)).toList();
@@ -49,22 +60,31 @@ public class CommandeController {
             return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
         }
     }
-
+/*
+Les commandes sont récupérées par un websocket
     @GetMapping("/liste")
     public ResponseEntity<List<CommandeListeProjection>> fetchAllCommandesListe() {
         List<CommandeListeProjection> commandes = commandeService.findAllCommandeListe();
         return new ResponseEntity<>(commandes, HttpStatus.OK);
     }
+*/
 
+/*
+Les commandes ne sont jamais récupéré par leur id
     @GetMapping("/{id}")
     public ResponseEntity<CommandeReadDTO> fetchCommandeById(@PathVariable  Long id) {
         Optional<Commande> commande = commandeService.findCommandeById(id);
         return commande.map(value -> new ResponseEntity<>(dtoTools.convertToDto(value, CommandeReadDTO.class), HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(null, HttpStatus.NOT_FOUND));
     }
+*/
 
-    @PostMapping
-    public ResponseEntity<?> createCommande(@RequestBody CommandeCreateDTO commandeDto) {
+    //Utilisé lors de l'envoie du panier dans l'écran de prise de commande
+    @PostMapping(produces = "application/json")
+    public ResponseEntity<?> createCommande(@Valid @RequestBody CommandeCreateDTO commandeDto) {
         try  {
+            if (!ControlInputTool.isValidObject(commandeDto, CommandeCreateDTO.class)) {
+                return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+            }
             Commande commande = commandeService.createCommande(dtoTools.convertToEntity(commandeDto, Commande.class));
             CommandeReadDTO commandeReadDTO = dtoTools.convertToDto(commande, CommandeReadDTO.class);
             this.template.convertAndSend("/socket/commande", commandeReadDTO);
@@ -75,14 +95,21 @@ public class CommandeController {
     }
 
 
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Boolean> deleteCommande(long id) {
+    //Utilisé lors de la suppression de des commandes dans l'écran de cuisine
+    @DeleteMapping(value = "/{id}", produces = "application/json")
+    public ResponseEntity<Boolean> deleteCommande(@PathVariable Long id) {
+        if (!ControlInputTool.isValidNumber(id)) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }        
         return new ResponseEntity<>(commandeService.deleteCommande(id), HttpStatus.OK);
     }
 
-    @PutMapping("/{id}")
+    //Utilisé lors notamment lors de l'envoie de commande dans l'écran de cuisine
+    @PutMapping(value = "/{id}", produces = "application/json")
     public ResponseEntity<CommandeReadDTO> updateCommande(@PathVariable long id) {
+        if (!ControlInputTool.isValidNumber(id)) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        } 
         Commande commande = commandeService.advanceStatusCommande(id);
         return new ResponseEntity<>(dtoTools.convertToDto(commande, CommandeReadDTO.class), HttpStatus.OK);
     }
